FEAT: Added TypeScript absolute imports
Setup TypeScript absolute imports to get rid of ../../../ and setup src as base. It is a little trickier to setup absolute imports for TS in NodeJS, so for details in future refer to this stackoverflow question. In short we need to setup baseUrl, paths and also install tsconfig-paths as devDependency and add 'ts-node': {'require': ['tsconfig-paths/register]} P.S: Replace the single quotes with double quotes for json ;)